# Optimise getting an individual's symptoms