--- a/src/tlo/methods/symptommanager.py
+++ b/src/tlo/methods/symptommanager.py
@@ -530,16 +530,9 @@ class SymptomManager(Module):
                 person_has = self.bsh.has(
                     [person_id], disease_module.name, first=True, columns=sy_columns
                 )
-                return [s for s in self.symptom_names if person_has[f"sy_{s}"]]
+                return [s for s in self.symptom_names if person_has[self.get_column_name_for_symptom(s)]]
             else:
-                symptom_cols = df.loc[
-                    person_id, [f"sy_{s}" for s in self.symptom_names]
-                ]
-                return (
-                    symptom_cols.index[symptom_cols > 0]
-                    .str.removeprefix("sy_")
-                    .to_list()
-                )
+                return [s for s in self.symptom_names if df.at[person_id, self.get_column_name_for_symptom(s)] > 0]
 
     def have_what(self, person_ids: Sequence[int]):
         """Find the set of symptoms for a list of person_ids.